Finalizado el metodo GET.
Se soluciono el error en los generos. Ahora muestra que generos tiene un libro.

+++ backend/CRUD-books/java/src/main/java/com/example/java/model/Publisher.java
@@ -12,6 +12,6 @@ public class Publisher {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id_publisher;
+    private int id_publisher;
     private String name;
 }
