+++ backend/CRUD-books/java/src/main/java/com/example/java/repositories/IAuthorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IAuthorRepository extends JpaRepository<Author, Long> {
+public interface IAuthorRepository extends JpaRepository<Author, Integer> {
 }
